Continue to cleanup
