+++ bridgebot/cards.py
@@ -6,38 +6,38 @@ from enums import Suits, Ranks
 class InvalidCardException(Exception):
     pass
 
+
 class InvalidRankException(InvalidCardException):
     pass
+
 
 class InvalidSuitException(InvalidCardException):
     pass
 
+
 class InvalidIndexException(Exception):
     pass
+
 
 class CardNotInHandException(Exception):
     pass
 
+
 class CardDoesntFollowSuitException(Exception):
     pass
+
 
 class WrongSizeHandException(Exception):
     pass
 
 
-class Deck:
-    def __init__(self):
-        self.card_indices = list(range(52))
-
-    def shuffle(self):
-        for i in range(51, 0, -1):
-            j = random.randint(0, i)  # So as not to bias certain permutations
-            if j == i:
-                continue
-            self.card_indices[i], self.card_indices[j] = self.card_indices[j], self.card_indices[i]
-
-
 class Card:
+    @staticmethod
+    def generate_card_from_index(index):
+        if index not in range(52):
+            raise InvalidIndexException("Index must be an integer between 0 and 51 inclusive.")
+        return Card(Suits.suits()[int(index / 13)], Ranks.ranks()[index % 13])
+
     def __init__(self, suit, rank):
         if not isinstance(suit, Suits):
             raise InvalidSuitException("Suit not found")
@@ -52,16 +52,12 @@ class Card:
         return self.suit == other.suit and self.rank == other.rank
 
 
-def map_index_to_card(index):
-    if index not in range(52):
-        raise InvalidIndexException("Index must be an integer between 0 and 51 inclusive.")
-    return Card(Suits.suits()[int(index / 13)], Ranks.ranks()[index % 13])
-
 class DeckListNotValid(Exception):
     pass
 
-class Hand:
-    def __init__(self):
+
+class BridgeHand:
+    def __init__(self, deck_indices):
         self.hand = {
             Suits.SPADES: set(),
             Suits.HEARTS:  set(),
@@ -69,8 +65,18 @@ class Hand:
             Suits.CLUBS: set()
         }
 
+        if len(deck_indices) != 13:
+            raise WrongSizeHandException("Bridge hands must contain 13 cards.")
+
+        if not isinstance(deck_indices, list):
+            raise DeckListNotValid("deck_indices is not a list")
+
+        for index in deck_indices:
+            card = Card.generate_card_from_index(index)
+            self.__add_card(card)
+
     @staticmethod
-    def _check_input(card):
+    def __check_input(card):
         if not isinstance(card.suit, Suits):
             raise InvalidSuitException("Suit not found")
 
@@ -78,44 +84,33 @@ class Hand:
             raise InvalidRankException("Rank not found")
 
     def contains_card(self, card):
-        self._check_input(card)
+        """
+        Returns if a card is in the hand or not
+
+        Parameters
+        ----------
+        card: Card
+            The card to be checked
+
+        Returns
+        -------
+        card_in_hand: bool
+            Whether the card is in the hand or not
+
+        """
+        self.__check_input(card)
         return card.rank in self.hand[card.suit]
 
-    def play_card(self, card):
-        self._check_input(card)
+    def __play_card(self, card):
+        self.__check_input(card)
         if not self.contains_card(card):
             raise CardNotInHandException("Hand does not contain " + card.rank.name + " of " + card.suit.name + ".")
         self.hand[card.suit].difference_update([card.rank])
 
-    def add_card(self, card):
+    def __add_card(self, card):
         if not isinstance(card, Card):
             raise InvalidCardException("Invalid Card")
         self.hand[card.suit].add(card.rank)
-
-    # Take a number from 0 to 51 and map it to suit and rank.
-    def add_card_from_deck_index(self, index):
-        if index not in range(52):
-            raise InvalidIndexException("Index must be an integer between 0 and 51 inclusive.")
-        card = map_index_to_card(index)
-        self.add_card(card)
-
-
-
-    # Take a list of numbers from 0 to 51 and map them to suits and ranks.
-    def fill_from_list(self, deck_indices):
-        if not isinstance(deck_indices, list):
-            raise DeckListNotValid("deck_indices is not a list")
-
-        for idx in deck_indices:
-            self.add_card_from_deck_index(idx)
-
-
-class BridgeHand(Hand):
-    def __init__(self, deck_indices):
-        super().__init__()
-        if len(deck_indices) != 13:
-            raise WrongSizeHandException("Bridge hands must contain 13 cards.")
-        self.fill_from_list(deck_indices)
 
     def lead(self, card):
         """
@@ -124,12 +119,12 @@ class BridgeHand(Hand):
         card: Card
             The Card that was played
         """
-        self._check_input(card)
+        self.__check_input(card)
 
         if not self.contains_card(card):
             raise CardNotInHandException("Hand does not contain " + card.rank.name + " of " + card.suit.name + ".")
 
-        self.play_card(card)
+        self.__play_card(card)
 
     def follow(self, led_suit, card_played):
         """
@@ -144,13 +139,13 @@ class BridgeHand(Hand):
         if not isinstance(led_suit, Suits):
             raise InvalidSuitException("Invalid trump_suit")
 
-        self._check_input(card_played)
+        self.__check_input(card_played)
 
         if card_played.suit != led_suit:
-            if len(self.hand.suit) != 0:
+            if len(self.hand[led_suit]) != 0:
                 raise CardDoesntFollowSuitException("Must follow suit if possible.")
             else:
-                self.play_card(card_played)
+                self.__play_card(card_played)
 
     def legal_cards(self, led_suit=None):
         """
